refactor(pd): extract PathwaysPDSchedulerMixin from scheduler.py
Per review: move _pd_* helpers and P/D worker init into
disaggregation/pathways_scheduler.py so scheduler.py only carries the
wiring hooks (mesh selection, tp_worker init dispatch, get_next_batch
dispatch, gc gate). scheduler.py: 3033 -> 2598 lines.
